refactor: move validators out
